# openapi/LupiraMtgApi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Lupira MTG API",
    "description": "Backend for the Lupira MTG mobile app — Magic: The Gathering card metadata, scan-based recognition, and per-user collection management. Authenticate with a Bearer token issued by Authentik (OIDC).",
    "version": "v1"
  },
  "paths": {
...
    "/scans": {
      "post": {
        "tags": [
          "Scans"
        ],
        "summary": "Recognize a Magic card from a photo.",
        "description": "Multipart upload — send the photo as a `multipart/form-data` field named `image`\n(jpeg, ≤4 MB). The backend runs perceptual-hash matching against the local Scryfall\nart-crop index AND OCR via FlorenceApi in parallel, then fuses the two signals into\na confidence-ranked list of `CardCandidate` objects.\n\nThe `confidence` field (`high|medium|low`) drives the mobile UX: HIGH auto-adds the\ntop candidate to the active selection; MEDIUM offers a disambiguation picker; LOW\nfalls back to manual lookup.",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "image"
                ],
                "type": "object",
                "properties": {
                  "image": {
                    "$ref": "#/components/schemas/IFormFile"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/scans/{scanId}/feedback": {
      "post": {
        "tags": [
          "Scans"
        ],
        "summary": "Report the actual correct printing for a previous scan.",
        "description": "Tells the API which printing the user actually wanted from a scan referenced\nby `scanId` (returned in `POST /scans`). The response reports the 1-based\n`rank` of that printing in the original candidate list, or `null` when the\nprinting wasn't in the pool at all.\n\nThe submission is persisted on the scan record and accumulates as training\ndata for future ranker work. Submitting a second time overwrites the prior\nfeedback — assume the latest is the user's most considered answer.",
        "parameters": [
          {
            "name": "scanId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScanFeedbackRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanFeedbackResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
...
    "/me/scans": {
      "get": {
        "tags": [
          "Me"
        ],
        "summary": "Paginated scan history (newest first).",
        "description": "Returns the caller's recent scans with the top match hydrated for each. `take` is\n1–100 (default 25); `skip` for paging. Use `GET /me/scans/{scanId}` for the full\nScanLog projection (every candidate, OCR zones, set-symbol detection, feedback).",
        "parameters": [
          {
            "name": "take",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/me/scans/{scanId}": {
      "get": {
        "tags": [
          "Me"
        ],
        "summary": "Full detail of one scan.",
        "description": "Returns every candidate (hydrated with the printing's metadata + presigned image),\nthe OCR zone texts, the set-symbol detection if any, the originally captured image\n(presigned URL, 15-min TTL) when retained, and feedback if submitted. Useful for\na \"why did the scan match this card?\" diagnostic UI without needing OpenObserve.",
        "parameters": [
          {
            "name": "scanId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanDetailResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
...
    "/selections": {
      "post": {
        "tags": [
          "Selections"
        ],
        "summary": "Start a new scanning selection.",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SelectionResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/selections/{selectionId}": {
      "get": {
        "tags": [
          "Selections"
        ],
        "summary": "Get a selection with its current cards.",
        "parameters": [
          {
            "name": "selectionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SelectionResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/selections/{selectionId}/cards": {
      "post": {
        "tags": [
          "Selections"
        ],
        "summary": "Add a recognized card to the selection. Returns 409 on duplicate unless allowDuplicate=true.",
        "parameters": [
          {
            "name": "selectionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddSelectionEntryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SelectionEntryResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/selections/{selectionId}/cards/{instanceId}": {
      "delete": {
        "tags": [
          "Selections"
        ],
        "summary": "Remove a card from the selection.",
        "parameters": [
          {
            "name": "selectionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "instanceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/selections/{selectionId}/commit": {
      "post": {
        "tags": [
          "Selections"
        ],
        "summary": "Commit selection cards into a collection.",
        "parameters": [
          {
            "name": "selectionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommitSelectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommitSelectionResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "AddSelectionEntryRequest": {
        "required": [
          "printingId"
        ],
        "type": "object",
        "properties": {
          "printingId": {
            "type": "string"
          },
          "isFoil": {
            "type": "boolean"
          },
          "language": {
            "type": "string"
          },
          "condition": {
            "type": "string"
          },
          "confidence": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "allowDuplicate": {
            "type": "boolean"
          }
        }
      },
...
      "CardCandidateResponse": {
        "required": [
          "printing",
          "combinedScore",
          "ocrAggregateScore",
          "nameScore",
          "typeLineScore",
          "rulesTextScore",
          "powerToughnessScore",
          "bottomMetadataScore",
          "hammingScore",
          "setTypeWeight"
        ],
        "type": "object",
        "properties": {
          "printing": {
            "$ref": "#/components/schemas/CardPrintingResponse"
          },
          "combinedScore": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "ocrAggregateScore": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "nameScore": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "typeLineScore": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "rulesTextScore": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "powerToughnessScore": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "bottomMetadataScore": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "hammingScore": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "setTypeWeight": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "hammingDistance": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "matchedByPHash": {
            "type": "boolean"
          },
          "matchedByName": {
            "type": "boolean"
          }
        }
      },
...
      "CommitSelectionRequest": {
        "required": [
          "collectionId"
        ],
        "type": "object",
        "properties": {
          "collectionId": {
            "type": "string",
            "format": "uuid"
          },
          "instanceIds": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        }
      },
      "CommitSelectionResponse": {
        "required": [
          "collectionId",
          "collectionName",
          "addedCount",
          "remainingCount"
        ],
        "type": "object",
        "properties": {
          "collectionId": {
            "type": "string",
            "format": "uuid"
          },
          "collectionName": {
            "type": "string"
          },
          "addedCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "remainingCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
...
      "IFormFile": {
        "type": "string",
        "format": "binary"
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "RecognitionConfidence": {
        "enum": [
          "Low",
          "Medium",
          "High"
        ]
      },
...
      "ScanDebug": {
        "required": [
          "zones",
          "isCropped",
          "cropConfidence",
          "cropRotated",
          "rotationRetried",
          "croppedWidth",
          "croppedHeight",
          "ocrRegionCount",
          "pHashCandidateCount",
          "ocrCandidateCount",
          "ocrLatencyMs",
          "pHashLatencyMs"
        ],
        "type": "object",
        "properties": {
          "zones": {
            "$ref": "#/components/schemas/ScanZoneTexts"
          },
          "setSymbol": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ScanSetSymbol"
              }
            ]
          },
          "imagePHash": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "isCropped": {
            "type": "boolean"
          },
          "cropConfidence": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "cropRotated": {
            "type": "boolean"
          },
          "rotationRetried": {
            "type": "boolean"
          },
          "croppedWidth": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "croppedHeight": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "ocrRegionCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "pHashCandidateCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "ocrCandidateCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "ocrLatencyMs": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "pHashLatencyMs": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
      "ScanDetailResponse": {
        "required": [
          "id",
          "scannedAt",
          "confidence",
          "ocrZones",
          "candidates"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "scannedAt": {
            "type": "string",
            "format": "date-time"
          },
          "confidence": {
            "$ref": "#/components/schemas/RecognitionConfidence"
          },
          "imageUrl": {
            "type": [
              "null",
              "string"
            ]
          },
          "imageMediaType": {
            "type": [
              "null",
              "string"
            ]
          },
          "ocrZones": {
            "$ref": "#/components/schemas/ScanZoneTexts"
          },
          "setSymbol": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ScanSetSymbol"
              }
            ]
          },
          "candidates": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CardCandidateResponse"
            }
          },
          "feedback": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ScanFeedbackInfo"
              }
            ]
          }
        }
      },
      "ScanFeedbackInfo": {
        "required": [
          "correctPrintingId",
          "at"
        ],
        "type": "object",
        "properties": {
          "correctPrintingId": {
            "type": "string"
          },
          "correctPrintingRank": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ScanFeedbackRequest": {
        "required": [
          "correctPrintingId"
        ],
        "type": "object",
        "properties": {
          "correctPrintingId": {
            "type": "string"
          }
        }
      },
      "ScanFeedbackResponse": {
        "required": [
          "scanId",
          "correctPrintingId",
          "rank",
          "candidateCount"
        ],
        "type": "object",
        "properties": {
          "scanId": {
            "type": "string",
            "format": "uuid"
          },
          "correctPrintingId": {
            "type": "string"
          },
          "rank": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "candidateCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
      "ScanListResponse": {
        "required": [
          "results",
          "total"
        ],
        "type": "object",
        "properties": {
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ScanSummaryResponse"
            }
          },
          "total": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
      "ScanResponse": {
        "required": [
          "scanId",
          "confidence",
          "candidates",
          "debug"
        ],
        "type": "object",
        "properties": {
          "scanId": {
            "type": "string",
            "format": "uuid"
          },
          "confidence": {
            "$ref": "#/components/schemas/RecognitionConfidence"
          },
          "candidates": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CardCandidateResponse"
            }
          },
          "debug": {
            "$ref": "#/components/schemas/ScanDebug"
          }
        }
      },
      "ScanSetSymbol": {
        "required": [
          "setCode",
          "hammingDistance",
          "score"
        ],
        "type": "object",
        "properties": {
          "setCode": {
            "type": "string"
          },
          "hammingDistance": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "score": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          }
        }
      },
      "ScanSummaryResponse": {
        "required": [
          "id",
          "scannedAt",
          "confidence"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "scannedAt": {
            "type": "string",
            "format": "date-time"
          },
          "confidence": {
            "$ref": "#/components/schemas/RecognitionConfidence"
          },
          "topMatch": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/CardPrintingResponse"
              }
            ]
          },
          "hasFeedback": {
            "type": "boolean"
          },
          "feedbackChangedTop": {
            "type": "boolean"
          }
        }
      },
...
      "SelectionEntryResponse": {
        "required": [
          "instanceId",
          "printing",
          "isFoil",
          "language",
          "condition",
          "confidence"
        ],
        "type": "object",
        "properties": {
          "instanceId": {
            "type": "string",
            "format": "uuid"
          },
          "printing": {
            "$ref": "#/components/schemas/CardPrintingResponse"
          },
          "isFoil": {
            "type": "boolean"
          },
          "language": {
            "type": "string"
          },
          "condition": {
            "type": "string"
          },
          "confidence": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          }
        }
      },
      "SelectionResponse": {
        "required": [
          "id",
          "cards",
          "createdAt",
          "expiresAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "cards": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SelectionEntryResponse"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "PoC device token from `POST /me/register`. Send as `Authorization: Bearer lmtg_<token>`. Replaced by Authentik OIDC tokens in the future Path C migration (see plan).",
        "scheme": "bearer"
      }
    }
  },
...
}